#137 Support default pagination
For larger datasets it usually it is mandatory to provide pagination
parameters. Otherwise the server either breaks down or takes long to a
nswer. This is is not obvious when working as a user in the browser and
usually leads to exactly this problem. There should be the possiblity to
declare a default page size if none is set explictly.

new API:
DefaultQuerySpecDeserializer.setDefaultOffset(...)
DefaultQuerySpecDeserializer.setDefaultLimit(...)