#15 pagination link computation for JPA adapter
added computation of page links for repositories managed by the
JpaModule. Support in Katharsis was limited for this, which required
some further refactorings:

- QueryParams/QuerySpec serializer added to katharsis-core (from client)
- based on this JsonApiUrlBuilder introduced for each creation of urls
- introduced PagedResultList for repositories to notify in cae of paging
about the total number of resources, allowing the computation of
pagination links.
- PagedRepositoryBase that implements link computation logic
- support in katharsis-client to read page links
- testing within katharsis-jpa/core/client

currently the mechanism (PagedResultList) is jpa-only. But it could be
moved into katharsis-core and the logic of PagedRepositoryBase could be
moved into the repository adapters. Then there would be a easy generic
solution as requested by #23 (Understanding Pagination Implementation)