#23 Understanding Pagination Implementation
Support to compute pagination links. If a repository
returns a PagedResultList with a total (unpaged) result count, links
will be computed by Katharsis.

The implementation is currently limited to when QuerySpec is used 
(meaing the application is setup wit a QuerySpecDeserializer instead of
a QueryParamsBuilder). QueryParams can be supported if needed with a
minor effort (complete QueryParamsAdapter implementation and handle
pagination different from offset/limit).