#177 Properly support patch/post semantic in katharsis-client
introduced create(...) method next to save() method to perform an
explicit push.

This breaks the current (wrong) contract of just having a save. A method
setPushAlways(...) is provided on KatharisClient to enable the old
behavior. For the next release if will be enabled to maintain backward
compatibility, but then it will be switched to the proper behavior in a
future release.