#121 Remove the need for relationship repositories
Implemented the new classes:

- QuerySpecResourceRepositoryBase
- QuerySpecRelationshipRepositoryBase

They take care of a lot of basic boiler-plate. In many cases
QuerySpecRelationshipRepositoryBase already provides full
implementation. And QuerySpecResourceRepositoryBase needs a single
method to have a fully functional read-only repository!

QuerySpecRelationshipRepositoryBase makes use of the resources
repositories on both side to implement its own functionality. In order
to make this possible, JsonApiToMany.opposite and JsonApiToOne.opposite
was introduced. The name of the opposite field allows e.g. to implement
findManyTarget and findOneTarget by searching for entities in the target
resource repository. Additional, removals and setting of relationships
is done with reflection on the resources and saved with the
corresponding resource repository.

Simplified the signature of the new bulk relationship interface a bit
along the way to make those two base classes as simple as possible to
use.