#208 ReposistoryInformationBuilder support
information extraction from a repository implemented the same way as
done for resources (ResourceInfromationBuilder). In the context of #24
where actions need to be implemented this approach allows more
flexibility like the jaxrs integration extending the repository
information based on the JAXRS annotations (i.e. introduction of actions
and configuration of repository url path).