#229 Support to differentiate between patch and post on the server side
Since we cannot change the existing internface without backward
compatiblity, to new interfaces will be introduced:

- ResourceRepositoryV2
- RelationshipRepositoryV2

extending the current QuerySpecRepository, providing a new method
create(...) and also make use of the new ResourceList.

The current ResourceRepositoryBase and RelationshipRepositoryBase are
updated accordingly to also implement the new interface.

katharsis-client is updated accordingly and with this commit no longer
consider to be experimental.