#233 support to decorate jpa repositories (#234)
JpaModule-based repositories provide the entire functionality out-of-the-box without any need for further implementation. To customize, implement actions, etc. it must be possible to wrap the
repository. Other potential future out-of-the-box implementation next to the JPA one will need a similar mechanism.

A new mechanism is introduced to register decorators. The decorators implements the new V2 repository interfaces and can perform any changes they like. As such the mechanism is applicable to any of kind. Decorators can be registered trough the module api with a factory.