improve JAXRS / JSONAPI interoperability #340
According to the JSONAPI specification, returning a data field containing null is legal. This is especially useful if warnings should be added to the response, which cannot be done when returning a 204 response.
Enhanced the JsonApiResponseFilter to also support returning a null data field.
This feature can be activated via feature flag 'katharsis.config.null.data.response.enabled'.