Support inclusions of complex object graphs #364
- introduced lookup behavior in KatharsisPropreties to switch between
including only paths from the requested (root) type to support arbitrary
inclusions with the include[type] syntax. The former adhering to the
JSON API spec, the later providing a powerful extension and used in
Katharsis already for quite a while (and for this reason staying the
default).
- switched abort criteria in IncludeLookupSetter from "attribute
processed" to "attribute for given resource processed". The same
attribute can this way be populated for different resources at different
positions in object graph. With this complex object graphs will get
properly supported.